+++ app/controller/admin/homeNav.js
@@ -9,6 +9,7 @@ class ControllerClass extends Controller {
       where: { name: { [ctx.app.Sequelize.Op.like]: `%${ctx.query.keyword || ''}%` } },
       order: [['sort']]
     })
+    rows.map(v => v.setDataValue('full_icon', ctx.domainName + v.icon))
     ctx.body = {
       data: rows,
       message: '查询成功'
@@ -34,7 +35,7 @@ class ControllerClass extends Controller {
     const result = await ctx.helper.getFormData();
     if (!row) ctx.throw(200, '数据不存在');
     // 删除旧的缩略图
-    if(row.icon != result.icon){
+    if(result.icon && row.icon != result.icon){
       const path = 'app/public' + row.icon;
       if(fs.existsSync(path)){ // 判断文件是否存在
         fs.unlinkSync(path); // 删除文件
